requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pandas==1.3.2